feat(accumulation): transfer counts, preimage keys

+++ modules/protocol/src/main/scala/io/forge/jam/protocol/accumulation/AccumulationTypes.scala
@@ -134,7 +134,7 @@ object AccumulationOperand:
   */
 final case class PreimageKey(
     hash: Hash,
-    length: Int
+    length: Long
 )
 
 /** Preimage request state.
